+++ test/functional/test.js
@@ -422,6 +422,51 @@ describe('Declarative Onboarding Functional Test Suite', function performFunctio
         }));
     });
 
+    describe('Test Auth Settings', function testAuth() {
+        this.timeout(1000 * 60 * 30); // 30 minutes
+        let body;
+        let currentState;
+
+        before(() => {
+            const thisMachine = machines[2];
+            const bigipAddress = thisMachine.ip;
+            const auth = { username: thisMachine.adminUsername, password: thisMachine.adminPassword };
+            const bodyFile = `${BODIES}/auth.json`;
+            return new Promise((resolve, reject) => {
+                common.readFile(bodyFile)
+                    .then((fileRead) => {
+                        body = JSON.parse(fileRead);
+                    })
+                    .then(() => common.testRequest(body, `${common.hostname(bigipAddress, constants.PORT)}`
+                            + `${constants.DO_API}`, auth, constants.HTTP_ACCEPTED, 'POST'))
+                    .then(() => common.testGetStatus(30, 60 * 1000, bigipAddress, auth,
+                        constants.HTTP_SUCCESS))
+                    .then((response) => {
+                        currentState = response.currentConfig.Common;
+                        resolve();
+                    })
+                    .catch((error) => {
+                        console.log(error);
+                        return common.dumpDeclaration(bigipAddress, auth);
+                    })
+                    .then((declarationStatus) => {
+                        reject(new Error(JSON.stringify(declarationStatus, null, 2)));
+                    })
+                    .catch((err) => {
+                        reject(err);
+                    });
+            });
+        });
+
+        it('should configure main auth settings', () => {
+            assert.ok(testMainAuth(body.Common.myAuth, currentState));
+        });
+
+        it('should configure radius', () => {
+            assert.ok(testRadiusAuth(body.Common.myAuth.radius, currentState));
+        });
+    });
+
     describe('Test Rollbacking', function testRollbacking() {
         this.timeout(1000 * 60 * 30); // 30 minutes
         let body;
@@ -647,6 +692,26 @@ function testConfigSyncIp(target, response) {
     const validRef = target.myConfigSync.configsyncIp === '/Common/mySelfIp/address';
     const validAddr = target.mySelfIp.address.indexOf(response.ConfigSync.configsyncIp) === 0;
     return validRef && validAddr;
+}
+
+function testMainAuth(target, response) {
+    return compareSimple(target, response.Authentication, ['enabledSourceType', 'fallback']);
+}
+
+function testRadiusAuth(target, response) {
+    const radiusResp = response.Authentication.radius;
+    const serviceTypeCheck = compareSimple(target, radiusResp, ['serviceType']);
+    const serverPrimaryCheck = compareSimple(
+        target.servers.primary,
+        radiusResp.servers.primary,
+        ['server', 'port']
+    );
+    const serverSecondaryCheck = compareSimple(
+        target.servers.secondary,
+        radiusResp.servers.secondary,
+        ['server', 'port']
+    );
+    return serviceTypeCheck && serverPrimaryCheck && serverSecondaryCheck;
 }
 
 /**